Time: 47 ms (50.53%) | Memory: 64.9 MB (74.53%) - LeetSync

+++ 2392-successful-pairs-of-spells-and-potions/successful-pairs-of-spells-and-potions.java
@@ -14,7 +14,7 @@ class Solution {
         int num = potions.length;
         while(l <= r){
             int mid = l + ((r - l) / 2);
-            long product = (long)((long)spell * (long)potions[mid]);
+            long product = (long) spell * potions[mid];
             if(product < target){
                 l = mid + 1;
             }
